requirements.txt
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cloudinary==1.30.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.1.2
django-allauth==0.51.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-countries==7.4.2
django-debug-toolbar==3.7.0
django-etc==1.4.0
django-hitcount==1.3.5
django-jazzmin==2.5.0
django-phonenumber-field==7.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-nested-routers==0.93.4
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.2
packaging==21.3
phonenumbers==8.12.57
Pillow==9.2.0
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.5
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
typing_extensions==4.4.0
tzdata==2022.5
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
